refactor: 테스트 메소드 패키지 변경
