+++ src/test/java/nextstep/ladder/model/GameResultTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
