+++ src/test/java/nextstep/ladder/model/PlayerTest.java
@@ -1,9 +1,8 @@
-package nextstep.ladder;
+package nextstep.ladder.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import nextstep.ladder.model.Player;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
