+++ src/test/java/nextstep/ladder/model/PlayersTest.java
@@ -1,11 +1,9 @@
-package nextstep.ladder;
+package nextstep.ladder.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 import java.util.List;
-import nextstep.ladder.model.Player;
-import nextstep.ladder.model.Players;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
